+++ general.h
@@ -10,7 +10,7 @@ float random_();
 float Gaussiana (float mu,float sigma);
 float norma(float x1,float y1,float x2, float y2, float L);
 float delta_x(float x1,float x2, float L);
-int imprimirMat(float *red, int n,int m);
+int imprimirMat(double *red, int n,int m);
